Re-organize file structure.
