--- src/auth/index.js
@@ -1,2 +1,0 @@
-import store from 'src/store'
-
