--- src/auth/reducer.js
@@ -1,3 +1,0 @@
-export default {
-  auth: state => state,
-}
