--- src/basic/hello.js
@@ -1,8 +1,0 @@
-const helloMessage = 'Hello from basic, again!'
-
-const sayHello = () => {
-  console.log('[hello]', helloMessage)
-  return helloMessage
-}
-
-export default sayHello
