+++ src/store/index.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware, compose } from 'redux'
 import { routerMiddleware } from 'react-router-redux'
 import createHistory from 'history/createBrowserHistory'
-import reducers from './reducers'
+import reducers from './reducer'
 
 const history = createHistory()
 const middleware = routerMiddleware(history)
